requirements.txt
anthropic==0.69.0
streamlit==1.50.0
pandas==2.3.3
numpy==2.3.3
openpyxl==3.1.5
python-dotenv==1.1.1
Pillow==11.1.0
scipy==1.15.1
cryptography==44.0.0
python-magic==0.4.27
plotly==5.24.1
reportlab==4.2.5
jinja2==3.1.5
lxml==5.3.0
watchdog==6.0.0
opencv-python==4.10.0.84
mss==10.0.0
pygetwindow==0.0.9
pyobjc-framework-Quartz==10.3.1
